Implement proper custom admin configuration.
Previous custom admin site configuration only allowed for the
registration and discovery of the immediately available site.
This commit reconfigures the admin site sub-classing and configuration
for the django.contrib.admin module to implement auto-discovery
of django-package models as it does with the default implementation.

Models can be registered as usual in the app/admin.py file.

Configuring admin site variables to be done in the /admin.py file.